+++ views/4_trabel.html
@@ -20,7 +20,7 @@
     <script type="text/javascript"
         src="//dapi.kakao.com/v2/maps/sdk.js?appkey=ed99290201dd1375b0f252f3e34ceb2f"></script>
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-    <script src="front_js/serverPage.js"></script>
+    <script src="front_js/showInfo.js"></script>
 
     </script>
